feat: 将初始化器和反初始化器重构为SDK内部类并增强功能
