+++ src/ErisPulse/Core/module.py
@@ -72,7 +72,7 @@ class ModuleManager(ManagerBase):
             raise TypeError(error_msg)
             
         if not issubclass(module_class, BaseModule):
-            warn_msg = f"模块 {module_name} 的类 {module_class.__name__} 没有继承自BaseModule，但我们仍会继续尝试加载这个模块，但请注意这可能引发其他问题"
+            warn_msg = f"模块 {module_name} 的类 {module_class.__name__} 虽然没有继承自BaseModule，但我们仍会继续尝试加载这个模块"
             logger.warning(warn_msg)
             # error_msg = f"模块 {module_name} 的类 {module_class.__name__} 必须继承自BaseModule"
             # logger.error(error_msg)
